fix: change enableRemoteModule to .enable() for electron@14 (#72)
BREAKING CHANGE: This removes the `enableRemoteModule` option in favor of a new `.enable()` method for enabling the remote module on a particular WebContents. See https://github.com/electron/remote/blob/main/docs/migration-2.md for migration instructions.

This is an empty commit for forcing a release after a semantic-release
failure in fd928e5.